+++ collector/src/main/java/com/bitmovin/analytics/ads/AdPosition.kt
@@ -1,5 +1,7 @@
 package com.bitmovin.analytics.ads
 
+import com.bitmovin.analytics.api.ssai.SsaiAdPosition
+
 @Suppress("ktlint")
 enum class AdPosition(val position: String) {
     PRE("pre"),
@@ -8,5 +10,15 @@ enum class AdPosition(val position: String) {
 
     override fun toString(): String {
         return position
+    }
+
+    fun mapToSsaiAdPosition(): SsaiAdPosition {
+        if (this.position == MID.position) {
+            return SsaiAdPosition.MIDROLL
+        } else if (this.position == POST.position) {
+            return SsaiAdPosition.POSTROLL
+        }
+
+        return SsaiAdPosition.PREROLL
     }
 }